Move *.js to *.json as it is not real JavaScript
The structure is as of a nice JSON file, but it has comments. Those
are all moved to __title and __description elements where __title is
a human-readable string and __description is either a human-readable
string or an array of human-readable strings that explain the steps
involved in the example one by one.